+++ appinfo/routes.php
@@ -2,7 +2,7 @@
 /**
  * Create your routes in here. The name is the lowercase name of the controller
  * without the controller part, the stuff after the hash is the method.
- * e.g. page#index -> OCA\DuplicateFinderBis\Controller\PageController->index()
+ * e.g. page#index -> OCA\DuplicateFinder\Controller\PageController->index()
  *
  * The controller class has to be registered in the application.php file since
  * it's instantiated in there
@@ -10,7 +10,7 @@
 return [
     'routes' => [
 	   ['name' => 'page#index', 'url' => '/', 'verb' => 'GET'],
-       ['name' => 'page#do_echo', 'url' => '/echo', 'verb' => 'POST'],
-       ['name' => 'page#files', 'url' => '/files', 'verb' => 'GET'],
+     ['name' => 'page#files', 'url' => '/files', 'verb' => 'GET'],
+     ['name' => 'duplicate_api#list', 'url' => '/api/{apiVersion}/Duplicates', 'verb' => 'GET', 'requirements' => ['apiVersion' => 'v1']],
     ]
 ];