+++ tests/Integration/AppTest.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace OCA\DuplicateFinderBis\Tests\Integration\Controller;
+namespace OCA\DuplicateFinder\Tests\Integration\Controller;
 
 use OCP\AppFramework\App;
-use Test\TestCase;
+use PHPUnit\Framework\TestCase;
 
 
 /**
@@ -15,15 +15,15 @@ class AppTest extends TestCase {
 
     private $container;
 
-    public function setUp() {
+    public function setUp() : void {
         parent::setUp();
-        $app = new App('duplicatefinderbis');
+        $app = new App('duplicatefinder');
         $this->container = $app->getContainer();
     }
 
     public function testAppInstalled() {
         $appManager = $this->container->query('OCP\App\IAppManager');
-        $this->assertTrue($appManager->isInstalled('duplicatefinderbis'));
+        $this->assertTrue($appManager->isInstalled('duplicatefinder'));
     }
 
 }